feat: Enhance image function cost management with resolution-aware pricing
- Introduced a new interface `ICostAwareFunction` for image functions to compute supplemental Query Unit costs based on image resolution.
- Implemented `ImageCostHelper` to calculate additional costs based on pixel count for image processing functions.
- Updated existing image functions (e.g., `BlurImageFunction`, `ResizeImageFunction`, etc.) to implement `ICostAwareFunction` and utilize the new cost calculation.
- Modified `DatumComputeOptions` to include a maximum Query Units setting for queries.
- Enhanced documentation to reflect changes in function costs and resource governance.
- Added comprehensive unit tests for image cost calculations and function implementations.

+++ src/DatumIngest/Execution/QueryMeter.cs
@@ -41,7 +41,7 @@ public sealed class QueryMeter
     /// Records the cost of a function invocation. Thread-safe.
     /// </summary>
     /// <param name="cost">The Query Unit cost to add.</param>
-    public void Add(int cost)
+    public void Add(long cost)
     {
         Interlocked.Add(ref _functionQueryUnits, cost);
     }
